requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pymilvus==2.6.0
transformers==4.44.0
sentence-transformers==2.2.2
torch==2.1.0
networkx==3.1
numpy==1.26.0
python-multipart==0.0.6
aiofiles==23.1.0
requests==2.31.0
Pillow==10.0.0
neo4j==5.11.0